pyproject.toml
[project]
name = "tennisbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.6.4",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
